+++ src/sys/device/display/x11/surface/mod.rs
@@ -1,0 +1,15 @@
+// devela::sys::device::display::x11::surface
+
+mod base; // (XSurface), (XSurfaceStorage)
+mod cpu; // XCpuBuffer
+mod shm; // XShmBuffer, (XShmCaps)
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            base::*,
+            cpu::*,
+            shm::*,
+        };
+    }
+}